cleanup: move ArchivalMaybeBinaryData to legacymodel
This diff concludes most of the development work associated with
https://github.com/ooni/probe/issues/2531. We move the nearly unused
ArchivalMaybeBinaryData to the internal/legacy/legacymodel package
to clearly mark it deprecated.

Finishing removing this struct is a task that is documented by
https://github.com/ooni/probe/issues/2543.
