+++ internal/experiment/quicping/quicping.go
@@ -17,6 +17,7 @@ import (
 
 	_ "crypto/sha256"
 
+	"github.com/ooni/probe-cli/v3/internal/legacy/legacymodel"
 	"github.com/ooni/probe-cli/v3/internal/legacy/tracex"
 	"github.com/ooni/probe-cli/v3/internal/model"
 )
@@ -78,26 +79,26 @@ type TestKeys struct {
 
 // SinglePing is a result of a single ping operation.
 type SinglePing struct {
-	ConnIdDst string                         `json:"conn_id_dst"`
-	ConnIdSrc string                         `json:"conn_id_src"`
-	Failure   *string                        `json:"failure"`
-	Request   *model.ArchivalMaybeBinaryData `json:"request"`
-	T         float64                        `json:"t"`
-	Responses []*SinglePingResponse          `json:"responses"`
+	ConnIdDst string                               `json:"conn_id_dst"`
+	ConnIdSrc string                               `json:"conn_id_src"`
+	Failure   *string                              `json:"failure"`
+	Request   *legacymodel.ArchivalMaybeBinaryData `json:"request"`
+	T         float64                              `json:"t"`
+	Responses []*SinglePingResponse                `json:"responses"`
 }
 
 type SinglePingResponse struct {
-	Data              *model.ArchivalMaybeBinaryData `json:"response_data"`
-	Failure           *string                        `json:"failure"`
-	T                 float64                        `json:"t"`
-	SupportedVersions []uint32                       `json:"supported_versions"`
+	Data              *legacymodel.ArchivalMaybeBinaryData `json:"response_data"`
+	Failure           *string                              `json:"failure"`
+	T                 float64                              `json:"t"`
+	SupportedVersions []uint32                             `json:"supported_versions"`
 }
 
 // makeResponse is a utility function to create a SinglePingResponse
 func makeResponse(resp *responseInfo) *SinglePingResponse {
-	var data *model.ArchivalMaybeBinaryData
+	var data *legacymodel.ArchivalMaybeBinaryData
 	if resp.raw != nil {
-		data = &model.ArchivalMaybeBinaryData{Value: string(resp.raw)}
+		data = &legacymodel.ArchivalMaybeBinaryData{Value: string(resp.raw)}
 	}
 	return &SinglePingResponse{
 		Data:              data,
@@ -272,7 +273,7 @@ L:
 					ConnIdDst: req.dstID,
 					ConnIdSrc: req.srcID,
 					Failure:   tracex.NewFailure(req.err),
-					Request:   &model.ArchivalMaybeBinaryData{Value: string(req.raw)},
+					Request:   &legacymodel.ArchivalMaybeBinaryData{Value: string(req.raw)},
 					T:         req.t,
 				})
 				continue
@@ -312,7 +313,7 @@ L:
 				ConnIdDst: ping.request.dstID,
 				ConnIdSrc: ping.request.srcID,
 				Failure:   tracex.NewFailure(timeoutErr),
-				Request:   &model.ArchivalMaybeBinaryData{Value: string(ping.request.raw)},
+				Request:   &legacymodel.ArchivalMaybeBinaryData{Value: string(ping.request.raw)},
 				T:         ping.request.t,
 			})
 			continue
@@ -325,7 +326,7 @@ L:
 			ConnIdDst: ping.request.dstID,
 			ConnIdSrc: ping.request.srcID,
 			Failure:   nil,
-			Request:   &model.ArchivalMaybeBinaryData{Value: string(ping.request.raw)},
+			Request:   &legacymodel.ArchivalMaybeBinaryData{Value: string(ping.request.raw)},
 			T:         ping.request.t,
 			Responses: responses,
 		})
