+++ internal/legacy/legacymodel/doc.go
@@ -1,0 +1,2 @@
+// Package legacymodel contains legacy content that used to be in internal/model
+package legacymodel
