+++ internal/model/archival_test.go
@@ -550,91 +550,6 @@ func TestArchivalScrubbedMaybeBinaryString(t *testing.T) {
 	})
 }
 
-func TestMaybeBinaryValue(t *testing.T) {
-	t.Run("MarshalJSON", func(t *testing.T) {
-		tests := []struct {
-			name    string // test name
-			input   string // value to marshal
-			want    []byte // expected result
-			wantErr bool   // whether we expect an error
-		}{{
-			name:    "with string input",
-			input:   "antani",
-			want:    []byte(`"antani"`),
-			wantErr: false,
-		}, {
-			name:    "with binary input",
-			input:   string(archivalBinaryInput),
-			want:    archivalEncodedBinaryInput,
-			wantErr: false,
-		}}
-		for _, tt := range tests {
-			t.Run(tt.name, func(t *testing.T) {
-				hb := model.ArchivalMaybeBinaryData{
-					Value: tt.input,
-				}
-				got, err := hb.MarshalJSON()
-				if (err != nil) != tt.wantErr {
-					t.Fatalf("ArchivalMaybeBinaryData.MarshalJSON() error = %v, wantErr %v", err, tt.wantErr)
-				}
-				if diff := cmp.Diff(tt.want, got); diff != "" {
-					t.Fatal(diff)
-				}
-			})
-		}
-	})
-
-	t.Run("UnmarshalJSON", func(t *testing.T) {
-		tests := []struct {
-			name    string // test name
-			input   []byte // value to unmarshal
-			want    string // expected result
-			wantErr bool   // whether we want an error
-		}{{
-			name:    "with string input",
-			input:   []byte(`"xo"`),
-			want:    "xo",
-			wantErr: false,
-		}, {
-			name:    "with nil input",
-			input:   nil,
-			want:    "",
-			wantErr: true,
-		}, {
-			name:    "with missing/invalid format",
-			input:   []byte(`{"format": "foo"}`),
-			want:    "",
-			wantErr: true,
-		}, {
-			name:    "with missing data",
-			input:   []byte(`{"format": "base64"}`),
-			want:    "",
-			wantErr: true,
-		}, {
-			name:    "with invalid base64 data",
-			input:   []byte(`{"format": "base64", "data": "x"}`),
-			want:    "",
-			wantErr: true,
-		}, {
-			name:    "with valid base64 data",
-			input:   archivalEncodedBinaryInput,
-			want:    string(archivalBinaryInput),
-			wantErr: false,
-		}}
-		for _, tt := range tests {
-			t.Run(tt.name, func(t *testing.T) {
-				hb := &model.ArchivalMaybeBinaryData{}
-				if err := hb.UnmarshalJSON(tt.input); (err != nil) != tt.wantErr {
-					t.Fatalf("ArchivalMaybeBinaryData.UnmarshalJSON() error = %v, wantErr %v", err, tt.wantErr)
-				}
-				if d := cmp.Diff(tt.want, hb.Value); d != "" {
-					t.Fatal(d)
-				}
-			})
-		}
-	})
-}
-
 func TestArchivalHTTPHeader(t *testing.T) {
 	t.Run("MarshalJSON", func(t *testing.T) {
 		tests := []struct {
